Move AsyncTestServer one step up
